+++ src/zrb/builtin/project/add/fastapp/fastapp_template/my_app_name/module/gateway/view/static/common/util.js
@@ -212,4 +212,12 @@ const UTIL = {
         return data;
     },
 
+    tryParseJSON(value) {
+        try {
+            return JSON.parse(value);
+        } catch {
+            return value;
+        }
+    },
+
 };